api/requirements.txt
Django==1.8.5
django-cors-headers==1.1.0
django-debug-toolbar==1.4
django-rest-swagger==0.3.4
djangorestframework==3.2.4
drf-nested-routers==0.10.0
fake-factory==0.5.3
pytz==2015.7
PyYAML==3.11
sqlparse==0.1.18
wheel==0.24.0
